Make rules a folder in template
As more circuit template rules will be added, having rules in a folder inside of template
will allow for better code organization.
